+++ components/Date/DateHeader.js
@@ -1,7 +1,18 @@
+// Import necessary modules from react, react-native and @expo/vector-icons
 import { StyleSheet, Text, View, TouchableOpacity } from "react-native";
 import React from "react";
 import { Ionicons } from "@expo/vector-icons";
 
+/**
+ * DateHeader is a functional component that renders a header with an icon, text, and an optional button.
+ * @param {Object} props - The props object.
+ * @param {boolean} props.isAllDaySelected - The selection state of the button.
+ * @param {Function} props.handleSubmit - The function to call when the button is pressed.
+ * @param {string} props.text - The text to display next to the icon.
+ * @param {JSX.Element} props.Icon - The icon to display.
+ * @param {boolean} props.showButton - Whether to show the button.
+ * @returns {JSX.Element} A View component with an icon, text, and an optional button.
+ */
 const DateHeader = ({
   isAllDaySelected,
   handleSubmit,
@@ -38,8 +49,10 @@ const DateHeader = ({
   );
 };
 
+// Export the DateHeader component as default
 export default DateHeader;
 
+// Define the styles for the dateTimeHeader, timeContainer, roundButton, and roundButtonSelected
 const styles = StyleSheet.create({
   dateTimeHeader: {
     flexDirection: "row",
