+++ components/Date/DateSelection.js
@@ -1,3 +1,4 @@
+// Import necessary modules from react, react-native and react-native-calendars
 import React, { useState } from "react";
 import {
   TouchableOpacity,
@@ -9,14 +10,24 @@ import {
 } from "react-native";
 import { Calendar } from "react-native-calendars";
 
+/**
+ * DateSelection is a functional component that renders a date picker.
+ * @param {Object} props - The props object.
+ * @param {Date} props.initialDate - The initial date to display.
+ * @param {Function} props.onDateChange - The function to call when the date changes.
+ * @returns {JSX.Element} A View component with a TouchableOpacity and a Modal.
+ */
 const DateSelection = ({ initialDate, onDateChange }) => {
+  // Define state variables for the visibility of the calendar and the selected date
   const [calendarShown, setCalendarShown] = useState(false);
   const [date, setDate] = useState(initialDate);
 
+  // Define a function to show the date picker
   const showDatePicker = () => {
     setCalendarShown(true);
   };
 
+  // Define a function to handle date changes
   const handleDateChange = (selectedDate) => {
     const [year, month, day] = selectedDate.split("-");
     const currentDate = new Date(year, month - 1, day);
@@ -25,6 +36,7 @@ const DateSelection = ({ initialDate, onDateChange }) => {
     onDateChange(currentDate);
   };
 
+  // Render a TouchableOpacity that shows the date picker when pressed, and a Modal that contains the calendar
   return (
     <>
       <View style={styles.border}>
@@ -52,8 +64,10 @@ const DateSelection = ({ initialDate, onDateChange }) => {
   );
 };
 
+// Export the DateSelection component as default
 export default DateSelection;
 
+// Define the styles for the centeredView, border, dateText, and modalView
 const styles = StyleSheet.create({
   centeredView: {
     flex: 1,
