+++ package.json
@@ -31,6 +31,7 @@
     "react-native-gesture-handler": "~2.14.0",
     "react-native-modal": "^13.0.1",
     "react-native-reanimated": "~3.6.2",
+    "react-native-vector-icons": "^10.0.3",
     "react-native-webview": "13.6.4"
   },
   "devDependencies": {
