+++ screens/AuthScreen/Auth.js
@@ -1,34 +1,46 @@
+// Import necessary modules from react, react-native, @expo/vector-icons, @react-navigation/native, expo-firebase-recaptcha, and firebase
 import React, { useState, useRef } from "react";
-import {
-  View,
-  TouchableOpacity,
-  TextInput,
-  Image,
-  StyleSheet,
-  Alert,
-} from "react-native";
-import Icon from "react-native-vector-icons/FontAwesome";
+import { View, TouchableOpacity, TextInput, Text, Image } from "react-native";
+import { Foundation } from "@expo/vector-icons";
+import Icon from "react-native-vector-icons/Ionicons";
+import { useNavigation } from "@react-navigation/native";
+
 import { FirebaseRecaptchaVerifierModal } from "expo-firebase-recaptcha";
-import { firebaseConfig } from "../../store/redux/config";
 import firebase from "firebase/compat/app";
 
+// Import local assets and styles
 import Logo from "../../assets/Logo.png";
 import ImageAsset from "../../assets/Image.png";
+import { firebaseConfig } from "../../store/redux/config";
+import styles from "./CSS/AuthStyle";
 
+/**
+ * Auth is a functional component that renders a phone number authentication form.
+ * @returns {JSX.Element} A View component with a form for phone number authentication.
+ */
 const Auth = () => {
+  // Define state variables for the phone number, verification code, and verification ID
+  const navigation = useNavigation();
   const [phoneNumber, setPhoneNumber] = useState("");
   const [code, setCode] = useState("");
   const [verificationId, setVerificationId] = useState(null);
   const recaptchaVerifier = useRef(null);
 
+  // Define a function to send a verification code to the phone number
   const sendVerification = async () => {
     const phoneProvider = new firebase.auth.PhoneAuthProvider();
-    phoneProvider
-      .verifyPhoneNumber(phoneNumber, recaptchaVerifier.current)
-      .then(setVerificationId);
-    setPhoneNumber("");
+    try {
+      const verificationId = await phoneProvider.verifyPhoneNumber(
+        phoneNumber,
+        recaptchaVerifier.current
+      );
+      setVerificationId(verificationId);
+    } catch (error) {
+      console.error(error);
+    }
   };
 
+  // Define a function to confirm the verification code and sign in the user
   const confirmCode = async () => {
     const credential = firebase.auth.PhoneAuthProvider.credential(
       verificationId,
@@ -39,14 +51,15 @@ const Auth = () => {
       .signInWithCredential(credential)
       .then(() => {
         setCode("");
+        navigation.navigate("Home");
       })
       .catch((error) => {
         console.log(error);
         alert("Invalid code");
       });
-    Alert.alert("Phone authentication successful 👍");
   };
 
+  // Render a form for phone number authentication
   return (
     <View style={styles.container}>
       <FirebaseRecaptchaVerifierModal
@@ -56,61 +69,67 @@ const Auth = () => {
       <Image source={Logo} style={styles.logo} />
       <Image source={ImageAsset} style={styles.image} />
       <View style={styles.inputContainer}>
-        <Icon name="mobile" size={20} color="#000" />
-        <TextInput
-          value={phoneNumber}
-          onChangeText={setPhoneNumber}
-          placeholder="Mobile No."
-          style={styles.input}
-          keyboardType="phone-pad"
-          autoCompleteType="tel"
-        />
+        <View style={styles.mobile}>
+          <Icon name="person-outline" size={20} color="#e7c2c2" />
+          <TextInput
+            value={phoneNumber}
+            onChangeText={setPhoneNumber}
+            placeholder="Mobile No."
+            style={{ ...styles.input, fontSize: 14 }}
+            keyboardType="phone-pad"
+            autoCompleteType="tel"
+            placeholderTextColor="#e7c2c2"
+          />
+          <View
+            style={{
+              borderLeftWidth: 1,
+              borderColor: "#9a9a9a",
+              height: 40,
+              marginHorizontal: 10,
+            }}
+          />
+          <TouchableOpacity onPress={() => sendVerification()}>
+            <Text style={styles.otp}>Send OTP</Text>
+          </TouchableOpacity>
+        </View>
       </View>
-      <TouchableOpacity onPress={() => sendVerification()}>
-        <Icon name="send" size={20} color="#000" />
-      </TouchableOpacity>
-      <TextInput
-        value={code}
-        onChangeText={setCode}
-        placeholder="Enter OTP"
-        style={styles.input}
-      />
-      <TouchableOpacity onPress={() => confirmCode()}>
-        <Icon name="check" size={20} color="#000" />
+
+      <View style={styles.otpContainer}>
+        <View
+          style={{
+            flexDirection: "row",
+            alignItems: "center",
+            paddingHorizontal: 15,
+          }}
+        >
+          <Foundation name="key" size={20} color="#e7c2c2" />
+          <TextInput
+            value={code}
+            onChangeText={setCode}
+            placeholder="Enter OTP"
+            style={{ ...styles.input, fontSize: 14 }}
+            placeholderTextColor="#e7c2c2"
+          />
+        </View>
+      </View>
+
+      <TouchableOpacity
+        onPress={() => confirmCode()}
+        style={{
+          backgroundColor: "blue",
+          width: "80%",
+          padding: 15,
+          alignItems: "center",
+          borderRadius: 10,
+          alignSelf: "center",
+          marginTop: 40,
+        }}
+      >
+        <Text style={{ color: "white" }}>Login</Text>
       </TouchableOpacity>
     </View>
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: "#fff",
-  },
-  logo: {
-    width: 100,
-    height: 50,
-    backgroundColor: "#fff",
-  },
-  image: {
-    width: 200,
-    height: 200,
-  },
-  inputContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  input: {
-    width: "80%",
-    padding: 10,
-    margin: 10,
-    borderWidth: 1,
-    borderColor: "#000",
-  },
-});
-
+// Export the Auth component as default
 export default Auth;
