+++ screens/HistoryScreen/History.js
@@ -1,28 +1,53 @@
-import { StyleSheet, Text, View } from "react-native";
+// Importing necessary modules
+import { Text, View } from "react-native";
 import React, { useContext } from "react";
 
+// Import local assets and styles
 import { EventContext } from "../../store/context/event-context";
+import styles from "./CSS/HistoryStyle";
 
+// History component to display the list of events
 const History = ({ route, navigation }) => {
+  // Using context to get the events
   const { events } = useContext(EventContext);
 
+  // If there are no events, display a message
   if (!events.length) {
     return <Text>No events scheduled</Text>;
   }
 
+  // Function to format the date
+  const formatDate = (date) => {
+    const startDate = new Date(date);
+    const weekdayShort = startDate.toLocaleDateString("en-US", {
+      weekday: "short",
+    });
+    const weekdayLong = startDate.toLocaleDateString("en-US", {
+      weekday: "long",
+    });
+    const day = startDate.toLocaleDateString("en-US", { day: "numeric" });
+    const year = startDate.toLocaleDateString("en-US", { year: "numeric" });
+    return { weekdayShort, weekdayLong, day, year };
+  };
+
+  // Function to format the time
+  const formatTime = (time) => {
+    const startTimeDate = new Date(time);
+    return startTimeDate.toLocaleTimeString("en-US", {
+      hour: "numeric",
+      minute: "numeric",
+      hour12: true,
+    });
+  };
+
+  // Render the list of events
   return (
     <>
       {events.map((event, index) => {
-        const startDate = new Date(event.selectedDate);
-        const weekdayShort = startDate.toLocaleDateString("en-US", {
-          weekday: "short",
-        });
-        const weekdayLong = startDate.toLocaleDateString("en-US", {
-          weekday: "long",
-        });
-        const day = startDate.toLocaleDateString("en-US", { day: "numeric" });
-        const year = startDate.toLocaleDateString("en-US", { year: "numeric" });
-        const startTimeDate = new Date(event.startTime);
+        const { weekdayShort, weekdayLong, day, year } = formatDate(
+          event.selectedDate
+        );
+        const startTime = formatTime(event.startTime);
         return (
           <View key={index}>
             <View style={styles.container}>
@@ -39,13 +64,7 @@ const History = ({ route, navigation }) => {
               </View>
             </View>
             <View style={styles.dateContainer}>
-              <Text style={styles.date}>
-                {startTimeDate.toLocaleTimeString("en-US", {
-                  hour: "numeric",
-                  minute: "numeric",
-                  hour12: true,
-                })}
-              </Text>
+              <Text style={styles.date}>{startTime}</Text>
             </View>
             <View style={styles.hr} />
           </View>
@@ -55,54 +74,5 @@ const History = ({ route, navigation }) => {
   );
 };
 
+// Exporting the History component
 export default History;
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: "row",
-    alignItems: "center",
-    gap: 10,
-    padding: 10,
-    backgroundColor: "white",
-  },
-  dateText: {
-    flexDirection: "column",
-    alignItems: "center",
-    gap: 5,
-  },
-  dayShort: {
-    fontSize: 14,
-    fontFamily: "open-sans-reg",
-    color: "gray",
-  },
-  peopleContainer: {
-    flexDirection: "column",
-    alignItems: "flex-start",
-    gap: 5,
-  },
-  peopleText: {
-    fontSize: 10,
-    fontFamily: "open-sans-reg",
-  },
-  details: {
-    fontSize: 10,
-    fontFamily: "open-sans-reg",
-    color: "black",
-  },
-  hr: {
-    borderBottomColor: "black",
-    borderBottomWidth: 1,
-    width: "100%",
-  },
-  dateContainer: {
-    alignItems: "flex-end",
-    paddingRight: 15,
-    backgroundColor: "white",
-  },
-  date: {
-    fontSize: 10,
-    fontFamily: "open-sans-reg",
-    color: "gray",
-    paddingBottom: 5,
-  },
-});
