+++ screens/NotificationsScreen/Notification.js
@@ -1,25 +1,35 @@
-import { TouchableOpacity, StyleSheet, Text, View } from "react-native";
+// Importing necessary modules and components
+import { TouchableOpacity, Text, View } from "react-native";
 import React, { useContext } from "react";
 
+// Import local assets and styles
 import { EventContext } from "../../store/context/event-context";
+import styles from "./CSS/NotificationStyles";
 
+// Notification component to display the list of notifications
 const Notification = () => {
+  // Using context to get the events
   const { events } = useContext(EventContext);
 
+  // If there are no events, display a message
   if (!events.length) {
     return <Text>No events scheduled</Text>;
   }
 
+  // Render the list of notifications
   return (
     <>
       {events.map((event, index) => {
+        // Convert selectedDate and startTime to Date objects
         const startDate = new Date(event.selectedDate);
         const startTime = new Date(event.startTime);
+        // Format the date
         const weekdayLong = startDate.toLocaleDateString("en-US", {
           weekday: "long",
         });
         const day = startDate.toLocaleDateString("en-US", { day: "numeric" });
         const year = startDate.toLocaleDateString("en-US", { year: "numeric" });
+        // Render each notification
         return (
           <View key={index}>
             <View style={styles.container}>
@@ -67,75 +77,5 @@ const Notification = () => {
   );
 };
 
+// Exporting the Notification component
 export default Notification;
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    paddingHorizontal: 10,
-    paddingVertical: 20,
-  },
-  miniContainer: {
-    flexDirection: "row",
-    gap: 10,
-  },
-  roundedView: {
-    backgroundColor: "lightpink",
-    borderRadius: 50,
-    width: 40,
-    height: 40,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  whiteText: {
-    color: "white",
-    fontSize: 30,
-  },
-  peopleText: {
-    fontSize: 12,
-    fontFamily: "open-sans-reg",
-    fontWeight: "bold",
-  },
-  dateText: {
-    fontSize: 10,
-    fontFamily: "open-sans-reg",
-  },
-  buttonContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-  },
-  declineButton: {
-    backgroundColor: "red",
-    padding: 6,
-    borderRadius: 20,
-    flex: 1,
-    marginRight: 5,
-    marginTop: 8,
-  },
-  acceptButton: {
-    backgroundColor: "blue",
-    padding: 6,
-    borderRadius: 20,
-    flex: 1,
-    marginRight: 5,
-    marginTop: 8,
-  },
-  buttonText: {
-    color: "white",
-    textAlign: "center",
-    fontSize: 8,
-    fontWeight: "bold",
-  },
-  timeText: {
-    fontSize: 10,
-    fontFamily: "open-sans-reg",
-    color: "gray",
-    textAlign: "right",
-  },
-  hr: {
-    borderBottomColor: "black",
-    borderBottomWidth: 1,
-    width: "100%",
-  },
-});
