--- store/firebase-context.js
@@ -1,35 +1,0 @@
-import React, { createContext, useState, useEffect } from "react";
-
-// Initialize Firebase
-const firebaseConfig = {
-  apiKey: "your-api-key",
-  authDomain: "your-auth-domain",
-  projectId: "your-project-id",
-  storageBucket: "your-storage-bucket",
-  messagingSenderId: "your-messaging-sender-id",
-  appId: "your-app-id",
-};
-
-if (!firebase.apps.length) {
-  firebase.initializeApp(firebaseConfig);
-} else {
-  firebase.app();
-}
-
-export const FirebaseContext = createContext();
-
-export const FirebaseContextProvider = (props) => {
-  const [currentUser, setCurrentUser] = useState(null);
-
-  useEffect(() => {
-    firebase.auth().onAuthStateChanged((user) => {
-      setCurrentUser(user);
-    });
-  }, []);
-
-  return (
-    <FirebaseContext.Provider value={{ currentUser, firebase }}>
-      {props.children}
-    </FirebaseContext.Provider>
-  );
-};
